Reconfigure flight control inputs to accept external sources
- Deleted inputdevices package so that inputs entirely received from LWJGLWorld

- Renamed ExternalFlightControlsStateManager to FlightControlsStateManager
