+++ src/com/chrisali/javaflightsim/simulation/SimulationRunner.java
@@ -32,7 +32,6 @@ import com.chrisali.javaflightsim.lwjgl.LWJGLWorld;
 import com.chrisali.javaflightsim.lwjgl.events.WindowClosedListener;
 import com.chrisali.javaflightsim.simulation.datatransfer.FlightData;
 import com.chrisali.javaflightsim.simulation.datatransfer.FlightDataListener;
-import com.chrisali.javaflightsim.simulation.flightcontrols.ExternalFlightControlsStateManager;
 import com.chrisali.javaflightsim.simulation.flightcontrols.FlightControlsState;
 import com.chrisali.javaflightsim.simulation.flightcontrols.FlightControlsStateManager;
 import com.chrisali.javaflightsim.simulation.integration.Integrate6DOFEquations;
@@ -54,7 +53,7 @@ public class SimulationRunner implements Runnable, WindowClosedListener {
 
 	private SimulationController simController;
 	
-	private ExternalFlightControlsStateManager flightControlsManager;
+	private FlightControlsStateManager flightControlsManager;
 	private Integrate6DOFEquations simulation;
 	private LWJGLWorld outTheWindow;
 	
@@ -86,7 +85,7 @@ public class SimulationRunner implements Runnable, WindowClosedListener {
 		configureSimulationTime();
 		
 		logger.debug("Initializing flight controls manager...");
-		flightControlsManager = new ExternalFlightControlsStateManager(simController, timeMS);
+		flightControlsManager = new FlightControlsStateManager(simController, timeMS);
 		
 		logger.debug("Initializing simulation...");
 		simulation = new Integrate6DOFEquations(flightControlsManager.getControlsState(), configuration);
