+++ src/com/chrisali/javaflightsim/simulation/flightcontrols/FlightControlsStateManager.java
@@ -19,7 +19,12 @@
  ******************************************************************************/
 package com.chrisali.javaflightsim.simulation.flightcontrols;
 
+import java.util.Collections;
 import java.util.EnumSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.logging.log4j.LogManager;
@@ -27,50 +32,37 @@ import org.apache.logging.log4j.Logger;
 
 import com.chrisali.javaflightsim.interfaces.SimulationController;
 import com.chrisali.javaflightsim.interfaces.Steppable;
+import com.chrisali.javaflightsim.simulation.datatransfer.InputData;
+import com.chrisali.javaflightsim.simulation.datatransfer.InputDataListener;
 import com.chrisali.javaflightsim.simulation.flightcontrols.analysis.AnalysisControlInput;
 import com.chrisali.javaflightsim.simulation.flightcontrols.analysis.AnalysisControls;
-import com.chrisali.javaflightsim.simulation.inputdevices.AbstractDevice;
-import com.chrisali.javaflightsim.simulation.inputdevices.Joystick;
-import com.chrisali.javaflightsim.simulation.inputdevices.JoystickVisitor;
-import com.chrisali.javaflightsim.simulation.inputdevices.Keyboard;
-import com.chrisali.javaflightsim.simulation.inputdevices.KeyboardVisitor;
-import com.chrisali.javaflightsim.simulation.inputdevices.Mouse;
-import com.chrisali.javaflightsim.simulation.inputdevices.MouseVisitor;
-import com.chrisali.javaflightsim.simulation.setup.ControlsConfiguration;
+import com.chrisali.javaflightsim.simulation.setup.KeyCommand;
 import com.chrisali.javaflightsim.simulation.setup.Options;
 import com.chrisali.javaflightsim.simulation.setup.SimulationConfiguration;
 import com.chrisali.javaflightsim.simulation.utilities.FileUtilities;
 
 /**
- * Handles flight controls actuated by human interface devices. Also contains 
- * {@link AnalysisControlInput} functionality when simulation in Analysis Mode
+ * Handles flight controls actuated by external human interface devices that provide their input values using 
+ * {@link InputDataListener}. Also contains {@link AnalysisControlInput} functionality when simulation runs 
+ * in Analysis Mode
  * 
  * @author Christopher Ali
  *
  */
-public class FlightControlsStateManager implements Steppable {
+public class FlightControlsStateManager implements Steppable, InputDataListener {
 
 	private static final Logger logger = LogManager.getLogger(FlightControlsStateManager.class);
 
 	private FlightControlsState controlsState;
-	
-	private EnumSet<Options> options;
-	
-	private AtomicInteger simTimeMS;
-	
-	private AbstractDevice hidController;
-	private Keyboard hidKeyboard;
 	private AnalysisControls analysisControls;
-
 	private ControlParameterActuator actuator;
-	
-	private JoystickVisitor joystickVisitor;
-    private KeyboardVisitor keyboardVisitor;
-    private MouseVisitor mouseVisitor;
-	
+
+	private EnumSet<Options> options;
+	private AtomicInteger simTimeMS;
+
 	public FlightControlsStateManager(SimulationController simController, AtomicInteger simTimeMS) {
 		logger.debug("Initializing flight controls...");
-		
+
 		SimEvents.init(simController);
 
 		SimulationConfiguration simConfig = simController.getConfiguration();
@@ -79,77 +71,81 @@ public class FlightControlsStateManager implements Steppable {
 		actuator = new FlightControlActuator(simConfig, controlsState);
 
 		this.simTimeMS = simTimeMS;
-		
-		ControlsConfiguration controlsConfig = FileUtilities.readControlsConfiguration();
+
 		analysisControls = FileUtilities.readAnalysisControls();
-		
 		if (analysisControls != null) {
 			logger.debug(analysisControls.getAnalysisInputs().size() + " analysis flight control inputs found:");
 			logger.debug(analysisControls.toString());
 		}
-		
-		// Use controllers for pilot in loop simulation if ANALYSIS_MODE not enabled 
-		if (!options.contains(Options.ANALYSIS_MODE)) {
-			if (options.contains(Options.USE_JOYSTICK)) {
-				logger.debug("Joystick controller selected");
-				hidController = new Joystick();
-				joystickVisitor = new JoystickVisitor(controlsConfig.getJoystickAssignments(), actuator);
-			}
-			else if (options.contains(Options.USE_MOUSE)){
-				logger.debug("Mouse controller selected");
-				hidController = new Mouse();
-				mouseVisitor = new MouseVisitor(controlsState, actuator);
-			}
-			
-			hidKeyboard = new Keyboard();
-			keyboardVisitor = new KeyboardVisitor(controlsConfig.getKeyboardAssignments(), actuator);
-		}
 	}
-	
+
 	@Override
 	public void step() {
 		try {
-			// if not running in analysis mode, controls and options are updated with pilot input
-			// otherwise, controls updated using generated doublets
-			if (!options.contains(Options.ANALYSIS_MODE)) {
-				if (hidController != null) 
-					hidController.collectControlDeviceValues(joystickVisitor != null ? joystickVisitor : mouseVisitor);
-		
-				if (hidKeyboard != null)
-					hidKeyboard.collectControlDeviceValues(keyboardVisitor);
-			} else {
+			if (options.contains(Options.ANALYSIS_MODE))
 				analysisControls.updateFlightControls(simTimeMS, actuator);
-			}
 			
 			limitControls(controlsState);
 		} catch (Exception e) {
 			logger.error("Flight controls encountered an error!", e);
 		}
 	}
-	
+
+	@Override
+	public void onInputDataReceived(InputData inputData) {
+		try {
+			List<KeyCommand> keyCommands = Collections.synchronizedList(inputData.getKeyCommands());
+			Map<FlightControl, Float> joystickInputs = Collections.synchronizedMap(inputData.getJoystickInputs());
+
+			synchronized (keyCommands) {
+				Iterator<KeyCommand> i = keyCommands.iterator();
+				while (i.hasNext()) {
+					actuator.handleParameterChange(i.next(), 1.0f);
+				}
+			}
+
+			Set<FlightControl> s = joystickInputs.keySet();
+			synchronized (joystickInputs) {
+				Iterator<FlightControl> i = s.iterator();
+				while (i.hasNext()) {
+					FlightControl axis = i.next();
+					actuator.handleParameterChange(axis, joystickInputs.get(axis));
+				}
+			}
+		} catch (Exception e) {
+			logger.error("Flight controls encountered an error!", e);
+		}
+	}
+
 	@Override
 	public boolean canStepNow(int timeMS) {
 		return timeMS % 1 == 0;
 	}
-		
-	public void setSimTimeMS(AtomicInteger simTimeMS) { this.simTimeMS = simTimeMS;	}
 
-	public AtomicInteger getSimTimeMS() { return simTimeMS;	}
-	
-	public FlightControlsState getControlsState() { return controlsState; }
+	public void setSimTimeMS(AtomicInteger simTimeMS) {
+		this.simTimeMS = simTimeMS;
+	}
+
+	public AtomicInteger getSimTimeMS() {
+		return simTimeMS;
+	}
+
+	public FlightControlsState getControlsState() {
+		return controlsState;
+	}
 
 	/**
-	 *  Limit control inputs to sensible deflection values based on the minimum and maximum values defined for 
-	 *  each member of {@link FlightControl}
-	 *  
-	 * @param controlsState 
+	 * Limit control inputs to sensible deflection values based on the minimum and
+	 * maximum values defined for each member of {@link FlightControl}
+	 * 
+	 * @param controlsState
 	 */
-	private void limitControls(FlightControlsState controlsState) {		
+	private void limitControls(FlightControlsState controlsState) {
 		for (FlightControl flc : FlightControl.values()) {
-            if (controlsState.get(flc) > flc.getMaximum())
-                controlsState.set(flc, flc.getMaximum());
-            else if (controlsState.get(flc) < flc.getMinimum())
-                controlsState.set(flc, flc.getMinimum());
-        }
+			if (controlsState.get(flc) > flc.getMaximum())
+				controlsState.set(flc, flc.getMaximum());
+			else if (controlsState.get(flc) < flc.getMinimum())
+				controlsState.set(flc, flc.getMinimum());
+		}
 	}
 }
