--- src/com/chrisali/javaflightsim/simulation/inputdevices/KeyboardVisitor.java
@@ -1,69 +1,0 @@
-/*******************************************************************************
- * Copyright (C) 2016-2018 Christopher Ali
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- * 
- *  If you have any questions about this project, you can visit
- *  the project's GitHub repository at: http://github.com/chris-ali/j6dof-flight-sim/
- ******************************************************************************/
-package com.chrisali.javaflightsim.simulation.inputdevices;
-
-import java.util.Map;
-
-import com.chrisali.javaflightsim.simulation.flightcontrols.ControlParameterActuator;
-import com.chrisali.javaflightsim.simulation.flightcontrols.FlightControlsState;
-import com.chrisali.javaflightsim.simulation.flightcontrols.FlightControlsStateManager;
-import com.chrisali.javaflightsim.simulation.setup.KeyCommand;
-
-import net.java.games.input.Component;
-import net.java.games.input.Controller;
-
-/**
- * Visitor that {@link FlightControlsStateManager} calls to poll keyboard data via jInput and send to a {@link ControlParameterActuator}
- * object, which then updates the {@link FlightControlsState} object as needed
- * 
- * @author Christopher
- *
- */
-public class KeyboardVisitor implements InputDeviceVisitor {
-
-	private Map<Integer, KeyCommand> keyboardAssignments;
-	private ControlParameterActuator actuator;
-
-	public KeyboardVisitor(Map<Integer, KeyCommand> keyboardAssignments, ControlParameterActuator actuator) {
-		this.keyboardAssignments = keyboardAssignments;
-		this.actuator = actuator;
-	}
-	
-	@Override
-	public boolean canHandleDevice(Controller device) {
-		return !keyboardAssignments.isEmpty();
-	}
-
-	@Override
-	public void handleDeviceInput(Controller device) {
-		if(!device.poll()) 
-			return;
-		
-		for (Component component: device.getComponents()) {
-			String componentName = component.getIdentifier().getName().toUpperCase();
-			KeyCommand command = keyboardAssignments.get(componentName);
-			
-			if(command != null)					
-				actuator.handleParameterChange(command, component.getPollData());
-							
-			continue;
-		}
-	}
-}
